+++ src/macos/PeripheralBase.mm
@@ -73,9 +73,9 @@ bool PeripheralBase::is_connected() {
 
 bool PeripheralBase::is_connectable() { return is_connectable_; }
 
-void PeripheralBase::unpair() {
-    throw Exception::OperationNotSupported();
-}
+bool PeripheralBase::is_paired() { throw Exception::OperationNotSupported(); }
+
+void PeripheralBase::unpair() { throw Exception::OperationNotSupported(); }
 
 std::vector<BluetoothService> PeripheralBase::services() {
     PeripheralBaseMacOS* internal = (__bridge PeripheralBaseMacOS*)opaque_internal_;